SOCINT-406 restructure to remove cyclic dependency so login gets permissions from CCSvc immediately
