+++ app/user_auth.py
@@ -7,10 +7,17 @@ from onelogin.saml2.utils import OneLogin_Saml2_Utils
 from functools import wraps
 from datetime import datetime
 from structlog import get_logger
+from user_context import get_name, is_logged_in, get_logged_in_user
+from access import load_permissions
+
 
 saml_bp = Blueprint('saml', __name__)
 
 logger = get_logger()
+
+"""
+Endpoints and functions for user authentication, i.e. login and logout, using SAML
+"""
 
 
 @saml_bp.route('/saml/metadata/')
@@ -76,7 +83,7 @@ def sls():
     Process return from IDP after signing out.
     """
     auth, req = do_auth()
-    request_id = get_from_session('LogoutRequestID')
+    request_id = _get_from_session('LogoutRequestID')
     timed_out = session.get('timed_out', None)
     url = auth.process_slo(request_id=request_id, delete_session_cb=lambda: session.clear())
     errors = auth.get_errors()
@@ -105,7 +112,7 @@ def acs():
     Process return from IDP after we sign-in.
     """
     auth, req = do_auth()
-    request_id = get_from_session('AuthNRequestID')
+    request_id = _get_from_session('AuthNRequestID')
     auth.process_response(request_id=request_id)
     errors = auth.get_errors()
     if not auth.is_authenticated():
@@ -115,7 +122,8 @@ def acs():
     if len(errors) == 0:
         store_in_session(auth)
         logger.info('Successful login for user ' + get_logged_in_user())
-        log_session_info(auth)
+        _log_session_info(auth)
+        load_permissions()
         name = get_name()
         welcome_name = name if name else get_logged_in_user()
         flash('Welcome <b>' + welcome_name + '</b>', 'info')
@@ -132,7 +140,7 @@ def acs():
     return redirect('/')
 
 
-def log_session_info(auth):
+def _log_session_info(auth):
     expiry_secs = auth.get_session_expiration()
     if expiry_secs:
         expiry_formatted = datetime.utcfromtimestamp(expiry_secs).strftime('%Y-%m-%d %H:%M:%S')
@@ -217,60 +225,8 @@ def store_in_session(auth):
     session['samlSessionIndex'] = auth.get_session_index()
 
 
-def get_from_session(key):
+def _get_from_session(key):
     value = None
     if key in session:
         value = session[key]
     return value
-
-
-def get_logged_in_user():
-    return session.get('samlNameId', 'nobody')
-
-
-def is_logged_in():
-    return 'samlNameId' in session and len(session['samlNameId']) > 0
-
-
-def get_attributes():
-    """
-    get the attributes dictionary from the session that was stored when we logged in.
-    """
-    attributes = None
-    if 'samlUserdata' in session:
-        if len(session['samlUserdata']) > 0:
-            attributes = session['samlUserdata']
-            logger.info('attributes: ' + str(attributes))
-    return attributes
-
-
-def get_name():
-    """
-    Create a "forename surname" name from the attributes
-    """
-    name = ''
-    attributes = get_attributes()
-    if attributes:
-        forename_key = 'givenname'
-        surname_key = 'surname'
-        if current_app.config['ADFS'] == 'True':
-            claim_prefix = 'http://schemas.xmlsoap.org/ws/2005/05/identity/claims/'
-            forename_key = claim_prefix + forename_key
-            surname_key = claim_prefix + surname_key
-
-        forename = session['samlUserdata'].get(forename_key, [None])[0]
-        surname = session['samlUserdata'].get(surname_key, [None])[0]
-        if forename:
-            name = forename
-        if surname:
-            name = name + ' ' + surname
-    return name
-
-
-def setup_auth_utilities(application):
-    """
-    Set up utility methods that can be called from the jinja2 HTML templates.
-    """
-    @application.context_processor
-    def utility_processor():
-        return dict(get_id=get_logged_in_user, is_logged_in=is_logged_in)
